+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## Unreleased
+
+* Added:
+  * Support for multiple backups per game. This requires Ludusavi 0.12.0 or newer.
+    You can open Ludusavi to configure how many backups to keep.
+    The feature will be disabled if an older version is detected.
+
 ## v0.9.0 (2022-08-13)
 
 * Added: